Delete validation results for deleted submissions.
